+++ receiver/receiver.h
@@ -1,6 +1,6 @@
 #ifndef RECEIVER_H
 #define RECEIVER_H
 
-void getMaxValue(float *maxValue, float newValue, int instanceNum);
+void getMaxValue(double *maxValue, double newValue, int instanceNum);
 
 #endif //RECEIVER_H
